KP-6210 Split custom_operators module
The custom_operators module was getting pretty hefty, and would grow
even more when the new access image downloading functionality is added.
The operators that download a single file of a certain type
(`SaveMetsSFTPOperator`, `SaveAltosSFTPOperator` and their parent class
`SaveFilesSFTPOperator`) are a set that clearly belongs together and are
somewhat disjoin from the more generic operators (e.g.
`CreateConnectionOperator`) and the operators with more wide scope
responsibilities (such as those that orchestrate downloading a bigger
batch of files). Thus they were selected to be placed in a module of
their own.
